登录登出完善新版

+++ my-project-backend/src/main/java/com/example/utils/JwtUtils.java
@@ -77,7 +77,7 @@ public class JwtUtils {
      * @param user 用户信息
      * @return 令牌
      */
-    public String createJwt(UserDetails user, String username, int userId) {
+    public String createJwt(UserDetails user, String username, Long userId) {
         if(this.frequencyCheck(userId)) {
             Algorithm algorithm = Algorithm.HMAC256(key);
             Date expire = this.expireTime();
@@ -146,7 +146,7 @@ public class JwtUtils {
      * @param userId 用户ID
      * @return 是否通过频率检测
      */
-    private boolean frequencyCheck(int userId){
+    private boolean frequencyCheck(Long userId){
         String key = Const.JWT_FREQUENCY + userId;
         return utils.limitOnceUpgradeCheck(key, limit_frequency, limit_base, limit_upgrade);
     }
